Layout seems to work now. Kind of.
However, it seems to fail on any nodes which do not have a full flower.
I will have to change the logic in the case where a node does not have a
full flower. This means adding some logic to the Node class as well as in
the layout routine.